Align tests with project conventions: extract helpers/mocks, fix function names, rename test file
