--- Tests/WalletTests/GetOrCreateTests.swift
@@ -1,226 +1,0 @@
-//
-//  GetOrCreateTests.swift
-//  CrossmintSDK
-//
-//  Created by Tomas Martins on 20/05/26.
-//
-
-import CrossmintCommonTypes
-import Foundation
-import SecureStorage
-import Testing
-@testable import Wallet
-
-private func makeTestWalletApiModel(address: String = "0x1234567890123456789012345678901234567890") -> WalletApiModel {
-    let json = """
-    {
-        "type": "smart",
-        "chainType": "evm",
-        "config": {
-            "adminSigner": {
-                "type": "email",
-                "email": "test@example.com",
-                "locator": "email:test@example.com"
-            }
-        },
-        "address": "\(address)",
-        "linkedUser": "email:test@example.com",
-        "createdAt": "2025-01-01T00:00:00.000Z"
-    }
-    """
-    let decoder = JSONDecoder()
-    decoder.dateDecodingStrategy = .iso8601
-    // swiftlint:disable:next force_try
-    return try! decoder.decode(WalletApiModel.self, from: Data(json.utf8))
-}
-
-@Suite("DefaultCrossmintWallets WalletSigner overloads", .tags(.unit))
-@MainActor struct GetOrCreateTests {
-
-    @Test("getWallet(signer:) returns existing wallet")
-    func getWalletReturnsExistingWallet() async throws {
-        let service = GetOrCreateMockWalletService()
-        service.getWalletResult = makeTestWalletApiModel(address: "0x1111111111111111111111111111111111111111")
-        let wallets = DefaultCrossmintWallets(
-            service: service,
-            secureWalletStorage: GetOrCreateMockSecureStorage()
-        )
-
-        let wallet = try await wallets.getWallet(
-            chain: Chain("base-sepolia"),
-            signer: .email("test@example.com") { _ in }
-        )
-
-        #expect(wallet?.address == "0x1111111111111111111111111111111111111111")
-        #expect(service.getWalletCallCount == 1)
-        #expect(service.createWalletCallCount == 0)
-    }
-
-    @Test("getWallet(signer:) returns nil when wallet not found")
-    func getWalletReturnsNilWhenNotFound() async throws {
-        let service = GetOrCreateMockWalletService()
-        service.getWalletResult = nil
-        let wallets = DefaultCrossmintWallets(
-            service: service,
-            secureWalletStorage: GetOrCreateMockSecureStorage()
-        )
-
-        let wallet = try await wallets.getWallet(
-            chain: Chain("base-sepolia"),
-            signer: .email("test@example.com") { _ in }
-        )
-
-        #expect(wallet == nil)
-        #expect(service.getWalletCallCount == 1)
-        #expect(service.createWalletCallCount == 0)
-    }
-
-    @Test("createWallet(signer:) creates and returns wallet")
-    func createWalletCreatesWallet() async throws {
-        let service = GetOrCreateMockWalletService()
-        service.createWalletResult = makeTestWalletApiModel(address: "0x2222222222222222222222222222222222222222")
-        let wallets = DefaultCrossmintWallets(
-            service: service,
-            secureWalletStorage: GetOrCreateMockSecureStorage()
-        )
-
-        let wallet = try await wallets.createWallet(
-            chain: Chain("base-sepolia"),
-            signer: .email("test@example.com") { _ in }
-        )
-
-        #expect(wallet.address == "0x2222222222222222222222222222222222222222")
-        #expect(service.getWalletCallCount == 0)
-        #expect(service.createWalletCallCount == 1)
-    }
-
-    @Test("Propagates error from getWallet")
-    func propagatesGetWalletError() async throws {
-        let service = GetOrCreateMockWalletService()
-        service.getWalletShouldThrow = true
-        let wallets = DefaultCrossmintWallets(
-            service: service,
-            secureWalletStorage: GetOrCreateMockSecureStorage()
-        )
-
-        await #expect(throws: WalletError.self) {
-            _ = try await wallets.getWallet(
-                chain: Chain("base-sepolia"),
-                signer: .email("test@example.com") { _ in }
-            )
-        }
-    }
-}
-
-// MARK: - Test doubles
-
-private final class GetOrCreateMockWalletService: SmartWalletService, @unchecked Sendable {
-    var authHeaders: [String: String] { [:] }
-    var isProductionEnvironment: Bool { false }
-
-    var getWalletResult: WalletApiModel?
-    var getWalletCallCount = 0
-    var getWalletShouldThrow = false
-
-    var createWalletResult: WalletApiModel?
-    var createWalletCallCount = 0
-
-    func getWallet(_ request: GetMeWalletRequest) async throws(WalletError) -> WalletApiModel {
-        getWalletCallCount += 1
-        if getWalletShouldThrow {
-            throw WalletError.walletGeneric("get wallet error")
-        }
-        guard let result = getWalletResult else {
-            throw WalletError.walletNotFound
-        }
-        return result
-    }
-
-    func createWallet(_ request: CreateWalletParams) async throws(WalletError) -> WalletApiModel {
-        createWalletCallCount += 1
-        guard let result = createWalletResult else {
-            throw WalletError.walletGeneric("create wallet result not configured")
-        }
-        return result
-    }
-
-    func getBalance(_ params: GetBalanceQueryParams) async throws(WalletError) -> Balances {
-        throw WalletError.walletGeneric("not implemented")
-    }
-
-    func getNFTs(_ params: GetNTFQueryParams) async throws(WalletError) -> [NFT] {
-        throw WalletError.walletGeneric("not implemented")
-    }
-
-    func createTransaction(
-        _ request: CreateTransactionRequest
-    ) async throws(TransactionError) -> any TransactionApiModel {
-        throw TransactionError.transactionGeneric("not implemented")
-    }
-
-    func signTransaction(_ request: SignRequest) async throws(TransactionError) -> any TransactionApiModel {
-        throw TransactionError.transactionGeneric("not implemented")
-    }
-
-    func fetchTransaction(
-        _ fetchTransactionRequest: FetchTransactionRequest
-    ) async throws(TransactionError) -> any TransactionApiModel {
-        throw TransactionError.transactionGeneric("not implemented")
-    }
-
-    func fund(_ request: FundWalletRequest) async throws(WalletError) {}
-
-    func transferToken(
-        _ request: TransferTokenRequest
-    ) async throws(TransactionError) -> any TransactionApiModel {
-        throw TransactionError.transactionGeneric("not implemented")
-    }
-
-    func createSignature(
-        _ request: CreateSignatureRequest
-    ) async throws(SignatureError) -> any SignatureApiModel {
-        throw SignatureError.unknown
-    }
-
-    func fetchSignature(
-        _ signatureId: String,
-        chainType: ChainType
-    ) async throws(SignatureError) -> any SignatureApiModel {
-        throw SignatureError.unknown
-    }
-
-    func approveSignature(_ request: SignRequest) async throws(SignatureError) {}
-
-    func removeSigner(
-        _ signerLocator: String,
-        chainType: ChainType,
-        chainName: String
-    ) async throws(TransactionError) -> any TransactionApiModel {
-        throw TransactionError.transactionGeneric("not implemented")
-    }
-
-    func listTransfers(_ params: ListTransfersQueryParams) async throws(WalletError) -> TransferListResult {
-        throw WalletError.walletGeneric("not implemented")
-    }
-
-    func addSigner(
-        _ entry: DelegatedSignerEntry,
-        chainType: ChainType,
-        chainName: String
-    ) async throws(WalletError) -> AddDelegatedSignerResponse {
-        AddDelegatedSignerResponse(chains: nil)
-    }
-
-    func registerTypedSigner(
-        _ signer: any AdminSignerData,
-        chainType: ChainType,
-        chainName: String
-    ) async throws(WalletError) -> AddDelegatedSignerResponse {
-        AddDelegatedSignerResponse(chains: nil)
-    }
-}
-
-private struct GetOrCreateMockSecureStorage: SecureWalletStorage {
-    func savePrivateKey(_ privateKey: String, forEmail email: String) {}
-    func getPrivateKey(forEmail email: String) -> String? { nil }
-}
